+++ examples/USBH_MIDI/bidrectional_converter/bidrectional_converter.ino
@@ -43,7 +43,7 @@ USB Usb;
 USBH_MIDI Midi(&Usb);
 
 void MIDI_poll();
-void doDelay(unsigned long t1, unsigned long t2, unsigned long delayTime);
+void doDelay(uint32_t t1, uint32_t t2, uint32_t delayTime);
 
 //If you want handle System Exclusive message, enable this #define otherwise comment out it.
 #define USBH_MIDI_SYSEX_ENABLE
@@ -149,9 +149,9 @@ void MIDI_poll()
 }
 
 // Delay time (max 16383 us)
-void doDelay(unsigned long t1, unsigned long t2, unsigned long delayTime)
+void doDelay(uint32_t t1, uint32_t t2, uint32_t delayTime)
 {
-  unsigned long t3;
+  uint32_t t3;
 
   if ( t1 > t2 ) {
     t3 = (0xFFFFFFFF - t1 + t2);
